% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCP.R
\name{CCP}
\alias{CCP}
\title{Consensus clustering}
\usage{
CCP(d = NULL, maxK = maxK, reps = reps, pItem = 0.8, pFeature = 1,
  clusterAlg = "hc", title = "untitled_consensus_cluster",
  innerLinkage = "average", finalLinkage = "average",
  distance = "euclidean", ml = NULL, tmyPal = NULL, seed = NULL,
  plot = "pdf", writeTable = TRUE, weightsItem = NULL,
  weightsFeature = NULL, verbose = F, corUse = "everything")
}
\arguments{
\item{d}{Numeric matrix. Data to be clustered, where columns=items/samples and rows are features}

\item{maxK}{Integer. Maximum number of clusters to evaluate}

\item{reps}{Integer. Number of subsamples so consensus can be evaluated}

\item{pItem}{Numerical value. Proportion of items (columns) to sample in each subsampling}

\item{pFeature}{Numerical value. Proportion of features (rows) to sample in each subsampling}

\item{clusterAlg}{Character string. Cluster algorithm: 'hc' heirarchical (hclust), 'pam' for paritioning around medoids, 'km' for k-means}

\item{title}{Character string. Name for output directory. Directory is created only if plot is not NULL or writeTable is TRUE. This title can be an abosulte or relative path.}

\item{innerLinkage}{heirarchical linkage method for subsampling}

\item{finalLinkage}{heirarchical linkage method for consensus matrix}

\item{distance}{Character string. 'pearson': (1 - Pearson correlation), 'spearman' (1 - Spearman correlation), 'euclidean', 'binary', 'maximum', 'canberra', 'minkowski" or custom distance function.}

\item{ml}{Optional. Prior result. If supplied then only do graphics and tables}

\item{tmyPal}{Optional. Character vector. Colors for consensus matrix}

\item{seed}{Optional. Numerical. Sets random seed for reproducible results}

\item{plot}{Character string. NULL - print to screen, 'pdf', 'png', 'pngBMP' for bitmap png, helpful for large datasets}

\item{writeTable}{Logical. TRUE - write ouput and log to csv}

\item{weightsItem}{Optional. Numerical vector. Weights to be used for sampling items}

\item{weightsFeature}{Optional. Numerical vector. Weights to be used for sampling features}

\item{verbose}{Logical. If TRUE, print messages to the screen to indicate progress. This is useful for large datasets}

\item{corUse}{Optional. Character string. Specifies how to handle missing data in correlation distances 'everything','pairwise.complete.obs', 'complete.obs'}
}
\value{
Returns a list including a lot of information for each k, most importantly the consensus matrices and class assignments
}
\description{
Takes an expression matrix and finds clusters via hierarchical clustering, k-means or partitioning around medoids. 
 Adapted from ConsensusClusterPlus of the ConsensusClusterPlus package, see \code{?ConsensusClusterPlus}.
}
\details{
Not intended for use outside of a call to \code{polyCluster}.
}
