% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonProp.R
\name{commonProp}
\alias{commonProp}
\title{Calculating the proportion of maximum intersection (PMI)}
\usage{
commonProp(l)
}
\arguments{
\item{l}{The output of a call to \code{performClust}.}
}
\value{
Returns an updated list \code{l}.
}
\description{
Finds the relative proportion of common samples between
 clusters from different algorithms, i.e. (A ∩ B)/min{|A|, |B|}.
 Then calculates the euclidean and cosine distance metrics for the propotions.
 Creates a pdf heatmap to show distances.
}
\details{
Not intended for use outside of a call to \code{polyCluster}.
}
