% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperClust2.R
\name{hyperClust2}
\alias{hyperClust2}
\title{Hypergeometric test of cluster overlap}
\usage{
hyperClust2(l)
}
\arguments{
\item{l}{The output of a call to \code{performClust}.}
}
\value{
Returns an updated list \code{l}.
}
\description{
Compares all clusters found by different algorithms via a
 hypergeometric test and creates a pdf heatmap to show
 the resulting p values.
 where n is the total number of clusters found across the algorithms.
}
\details{
Not intended for use outside of a call to \code{polyCluster}.
}
