% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputCond.R
\name{inputCond}
\alias{inputCond}
\title{Cutoff chooser}
\usage{
inputCond(init = 0, prompt = "Please enter a value for x: ",
  variable = "x", condition = "x < 3")
}
\arguments{
\item{init}{Numeric. The initial value of the variable}

\item{prompt}{Character string. Printed to prompt the user for input}

\item{variable}{Character string. The name of the variable being tested in the condition}

\item{condition}{Character string. The logical test of the variable. If true, the user input is returned.}
}
\value{
A user-entered value that meets the specified criteria.
}
\description{
Takes user input and tests it against a given condition.
}
\details{
Not intended for use outside of a call to \code{polyCluster}.
}
