% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeEdge.R
\name{nodeEdge}
\alias{nodeEdge}
\title{Node-edge graphs}
\usage{
nodeEdge(l, recons)
}
\arguments{
\item{l}{The output of a call to \code{performClust}.}

\item{recons}{Character vector. The names of the reconcilliation methods to be used (for naming/annotation purposes only)}
}
\value{
Returns an updated list \code{l}.
}
\description{
Creates a pdf node-edge graph from each reconciliation method
 and identifies which meta-clusters each sample belongs to. Then uses
 hyperClust3 to perform a hypergeometric test betwen meta-clusters
 and known phenotypes.
}
\details{
Not intended for use outside of a call to \code{polyCluster}.
}
