% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamCentroids.R
\name{pamCentroids}
\alias{pamCentroids}
\title{PAM centroids}
\usage{
pamCentroids(l)
}
\arguments{
\item{l}{The output of a call to \code{testClust}.}
}
\value{
Returns a list of PAM centroids for all communities.
}
\description{
Calculates PAM centroids of community subtypes.
}
\details{
Not intended for use outside of a call to \code{polyCluster}.
}
