% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.data.R
\name{read.gct}
\alias{read.gct}
\title{Read a text file of expression values}
\usage{
read.gct(filename = "NULL")
}
\arguments{
\item{filename}{A character vector given the file path of the gene expression data.}
}
\value{
Returns the expression dataset.
}
\description{
Reads a gene expression dataset with genes in rows, samples in columns
in .gct format and converts it into an R matrix
}
\details{
Not intended for use outside of a call to \code{polyCluster}.
}
