% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testClust.R
\name{testClust}
\alias{testClust}
\title{Test clustering result.}
\usage{
testClust(l)
}
\arguments{
\item{l}{List. The output of a call to performClust}
}
\value{
Returns an updated list \code{l}.
}
\description{
Compares clusters found by clustering algorithms via the proportion
 of common samples and a hypergeometric test.
}
\details{
Not intended for use outside of a call to \code{polyCluster}.
}
