% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyClustRPackage.R
\name{polyCluster}
\alias{polyCluster}
\title{polyCluster pipeline}
\usage{
polyCluster(filename, clusterAlg = c("hc", "pm", "km", "nmf"), maxK = 7,
  reps = 100, phenoFile = NULL, ref = "", nmfData = NULL,
  interactive = FALSE, seed = NULL)
}
\arguments{
\item{filename}{Character string. File name of preprocessed and normalized expression values with genes in rows and samples in columns}

\item{clusterAlg}{Character vector. Any combination of \code{"hc"} (hierarchical clustering), \code{"pm"} (partitioning around medoids), \code{"km"} (k-means) or \code{"nmf"} (nonnegative matrix factorization)}

\item{maxK}{Integer > 2. The maximum number of clusters of samples to evaluate.}

\item{reps}{Integer. The number of resampling (for \code{"hc"}, \code{"pm"} and \code{"km"}) or starting seed (\code{"nmf"}) iterations.}

\item{phenoFile}{Character string. File name of known phenotypes, with sample name in first column and phenotype in second column.}

\item{ref}{Character string. Reference with which to name the output of the analysis.}

\item{nmfData}{Character string. File path to the result of a previous NMF clustering by this function, in order to save time.}

\item{interactive}{Logical. If FALSE, networks will be plotted with an automatic layout. If TRUE, a users can customise the layout through tkplot.}

\item{seed}{Numeric. Value of seed to start clustering for reproducible results.}
}
\value{
Provides output from the whole pipeline in the current working directory.
}
\description{
Runs the entire polyCluster pipeline.
}
\examples{
library(polyClustR)

# Example breast dataset is provided with installation
exampleGE <- system.file('extdata', 'chin_breast_example_GE.txt', package = 'polyClustR')
exampleKnownSubtypes <- system.file('extdata', 'chin_breast_example_known_subtypes.txt', package = 'polyClustR')

polyCluster(exampleGE, clusterAlg = c('hc', 'km'), phenoFile = exampleKnownSubtypes, ref = 'test_run')
# Output is written to the current working directory.
}
